requirements.txt
requests==2.32.5
python-dotenv==1.2.1
tenacity==9.1.2
